Add files via upload
I added a top horizontal navigation bar on the landing page, all the links direct to the about html page for now. I also sorted all the files into folders so the references have changed. I have made a zip file that includes all other files in the proper directories as well. Use that to test out the site if need be. (1920x969 resolution don't forget)