Add files via upload
The website is now responsive to other screen resolutions, so no need to go into chrome and change the resolution anymore! Refreshing the page will fit it to the given screen width and scale the height to keep the aspect ratio as well. 

The about.html page has been updated with some drop downs. Slight visual glitch on the bottom tabs but that'll go away when I add more tabs to cover it up.