Add files via upload
Worked on the about.html page, which has been renamed to docs.html since about is a sub-page within it. Some jquery code has been added that will allow the links on the nav tab to change the content in the body of the page. An example about page from the Keras website has been taken for now. The about page is loaded by default and whenever the about tab is clicked on the side navigation bar. 

Minor CSS changes to fix the side bar to the page and push the page content to the side a bit. The site only uses pixel measurements for now but doesn't scale like the landing page, so it works best on a 1920x969 chrome browser.