Add files via upload
The jquery in the previous example gave a CORS (Cross-Origin Resource Sharing) error so the links for the html is now set to the raw file that can be found on this github. However, that means this repo must be public and quickly updating the code will be hard with having to push to github every time instead of working from a local file. 

TBD: fix it so somehow so that the innerhtml of the docs page can be loaded off a local file instead of a website link.