Add files via upload
Major changes to the API sites -> each tab now redirects to a new html page instead of loading in new content. Animations and CSS references were updated to work with the new page loading (although it looks slightly different). Tab links were added in to go from page to page, and space was added to add the page content into the html file later. 

I also updated the navigation bar to change its text so you can tell if the side navigation bar is working properly. It will be formatted properly in a future update.